# resources/docs/swagger.yaml
definitions:
...
  request.MenuCreateRequest:
    properties:
      component:
        description: 组件路径
        type: string
      icon:
        description: 菜单图标
        type: string
      isCache:
        description: 是否缓存（0缓存 1不缓存）
        type: string
      isFrame:
        description: 是否为外链（0是 1否）
        type: string
      menuKey:
        description: 菜单标识
        type: string
      menuName:
        description: 菜单名称
        type: string
      menuType:
        description: 菜单类型（D目录 M菜单 A按钮）
        enum:
        - D
        - M
        - A
        type: string
      orderNum:
        description: 显示顺序
        minimum: 0
        type: integer
      parentID:
        description: 父菜单ID
        minimum: 0
        type: integer
      path:
        description: 路由地址
        type: string
      remark:
        description: 备注
        type: string
      requestMethod:
        description: 请求方法
        enum:
        - GET
        - POST
        - DELETE
        - PUT
        type: string
      visible:
        description: 菜单状态（0显示 1隐藏）
        type: string
    required:
    - menuKey
    - menuName
    - path
    - requestMethod
    type: object
  request.MenuUpdateRequest:
    properties:
      component:
        description: 组件路径
        type: string
      icon:
        description: 菜单图标
        type: string
      id:
        description: 菜单ID
        type: integer
      isCache:
        description: 是否缓存（0缓存 1不缓存）
        type: string
      isFrame:
        description: 是否为外链（0是 1否）
        type: string
      menuKey:
        description: 菜单标识
        type: string
      menuName:
        description: 菜单名称
        type: string
      menuType:
        description: 菜单类型（D目录 M菜单 A按钮）
        enum:
        - D
        - M
        - A
        type: string
      orderNum:
        description: 显示顺序
        minimum: 0
        type: integer
      parentID:
        description: 父菜单ID
        minimum: 0
        type: integer
      path:
        description: 路由地址
        type: string
      remark:
        description: 备注
        type: string
      requestMethod:
        description: 请求方法
        enum:
        - GET
        - POST
        - DELETE
        - PUT
        type: string
      status:
        description: 状态（0正常 1停用）
        type: string
      visible:
        description: 菜单状态（0显示 1隐藏）
        type: string
    required:
    - id
    - menuKey
    - menuName
    - path
    - requestMethod
    type: object
...
  request.PostCreateRequest:
    properties:
      orderNum:
        description: 显示顺序
        minimum: 0
        type: integer
      postName:
        description: 岗位名称
        type: string
      remark:
        description: 备注
        type: string
    required:
    - postName
    type: object
  request.PostUpdateRequest:
    properties:
      id:
        description: 岗位ID
        type: integer
      orderNum:
        description: 显示顺序
        minimum: 0
        type: integer
      postName:
        description: 岗位名称
        type: string
      remark:
        description: 备注
        type: string
      status:
        description: 状态（0正常 1停用）
        type: string
    required:
    - id
    - postName
    type: object
...
  response.MenuResponse:
    properties:
      children:
        description: 子菜单
        items:
          $ref: '#/definitions/response.MenuResponse'
        type: array
      component:
        description: 组件路径
        type: string
      createBy:
        description: 创建者
        type: string
      createTime:
        description: 创建时间
        type: string
      delFlag:
        description: 删除标志（0代表存在 2代表删除）
        type: string
      icon:
        description: 菜单图标
        type: string
      id:
        description: 菜单ID
        type: integer
      isCache:
        description: 是否缓存（0缓存 1不缓存）
        type: integer
      isFrame:
        description: 是否为外链（0是 1否）
        type: integer
      menuKey:
        description: 菜单标识
        type: string
      menuName:
        description: 菜单名称
        type: string
      menuType:
        description: 菜单类型（D目录 M菜单 A按钮）
        type: string
      orderNum:
        description: 显示顺序
        type: integer
      parentID:
        description: 父菜单ID
        type: integer
      path:
        description: 路由地址
        type: string
      remark:
        description: 备注
        type: string
      requestMethod:
        description: 请求方法
        type: string
      status:
        description: 状态（0正常 1停用）
        type: string
      updateBy:
        description: 更新者
        type: string
      updateTime:
        description: 更新时间
        type: string
      visible:
        description: 菜单状态（0显示 1隐藏）
        type: string
    type: object
...
  response.PostResponse:
    properties:
      createBy:
        description: 创建者
        type: string
      createTime:
        description: 创建时间
        type: string
      delFlag:
        description: 删除标志（0代表存在 2代表删除）
        type: string
      id:
        description: 岗位ID
        type: integer
      orderNum:
        description: 显示顺序
        type: integer
      postName:
        description: 岗位名称
        type: string
      remark:
        description: 备注
        type: string
      status:
        description: 状态（0正常 1停用）
        type: string
      updateBy:
        description: 更新者
        type: string
      updateTime:
        description: 更新时间
        type: string
    type: object
  response.RestRes:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: status ok
        type: string
    type: object
...
  response.SessionResponse:
    properties:
      clientIp:
        description: 请求地址
        type: string
      createAt:
        description: 创建时间
        type: string
      expiresAt:
        description: 是否阻止
        type: string
      id:
        description: 唯一标识
        type: string
      isBlocked:
        description: 是否阻止
        type: boolean
      realName:
        description: 真实姓名
        type: string
      refreshToken:
        description: 刷新秘钥
        type: string
      remark:
        description: 备注
        type: string
      userAgent:
        description: 请求信息
        type: string
      userName:
        description: 用户名
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample golang project
  title: Admin GOGOGO
  version: 0.0.1
paths:
...
  /menu:
    post:
      consumes:
      - application/json
      description: 新增岗位信息
      parameters:
      - description: 新增岗位信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.MenuCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 岗位新增
      tags:
      - 系统菜单
    put:
      consumes:
      - application/json
      description: 更新岗位信息
      parameters:
      - description: 更新岗位信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.MenuUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 岗位更新
      tags:
      - 系统菜单
  /menu/:id:
    delete:
      consumes:
      - application/json
      description: 伪删除岗位信息
      parameters:
      - description: 标识
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 岗位伪删除
      tags:
      - 系统菜单
    get:
      consumes:
      - application/json
      description: 详情岗位信息
      parameters:
      - description: 标识
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.MenuResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 岗位详情
      tags:
      - 系统菜单
  /menu/list:
    get:
      consumes:
      - application/json
      description: 获取所有岗位信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.MenuResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 岗位列表查询
      tags:
      - 系统菜单
  /menu/p:
    post:
      consumes:
      - application/json
      description: 获取所有岗位信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.MenuResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 岗位列表查询
      tags:
      - 系统菜单
  /menu/tree:
    get:
      consumes:
      - application/json
      description: 获取所有菜单信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.MenuResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 菜单列表树
      tags:
      - 系统菜单
  /post:
    post:
      consumes:
      - application/json
      description: 新增岗位信息
      parameters:
      - description: 新增岗位信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.PostCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 岗位新增
      tags:
      - 系统岗位
    put:
      consumes:
      - application/json
      description: 更新岗位信息
      parameters:
      - description: 更新岗位信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.PostUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 岗位更新
      tags:
      - 系统岗位
  /post/:id:
    delete:
      consumes:
      - application/json
      description: 伪删除岗位信息
      parameters:
      - description: 标识
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RestRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 岗位伪删除
      tags:
      - 系统岗位
    get:
      consumes:
      - application/json
      description: 详情岗位信息
      parameters:
      - description: 标识
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.PostResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 岗位详情
      tags:
      - 系统岗位
  /post/list:
    get:
      consumes:
      - application/json
      description: 获取所有岗位信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.PostResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 岗位列表查询
      tags:
      - 系统岗位
  /post/p:
    post:
      consumes:
      - application/json
      description: 获取所有操作日志信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.RestRes'
            - properties:
                data:
                  $ref: '#/definitions/response.SessionResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RestRes'
      summary: 操作日志列表查询
      tags:
      - 系统操作日志
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"